# schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CTU Theseus API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/runs": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "List Runs",
        "operationId": "list_runs",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Get Run",
        "operationId": "get_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "training"
        ],
        "summary": "Delete Run",
        "description": "Delete a run whatever its status. An in-flight run is stopped first.",
        "operationId": "delete_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/train": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Start Training",
        "operationId": "start_training",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Cancel Run",
        "operationId": "cancel_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/status": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Get Run Status",
        "description": "Short-poll fallback for the live console.",
        "operationId": "get_run_status",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/evaluation": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Get Run Evaluation",
        "operationId": "get_run_evaluation",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/evaluation/errors": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Get Run Evaluation Errors",
        "description": "Paginated misclassified rows from the report, joined back to their pool items.",
        "operationId": "get_run_evaluation_errors",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "classId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Classid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationErrorsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "downloadUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Downloadurl"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text",
          "downloadUrl"
        ],
        "title": "ErrorItem"
      },
...
      "EvaluationErrorRow": {
        "properties": {
          "itemId": {
            "type": "string",
            "title": "Itemid"
          },
          "actual": {
            "type": "string",
            "title": "Actual"
          },
          "predicted": {
            "type": "string",
            "title": "Predicted"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "item": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorItem"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "itemId",
          "actual",
          "predicted",
          "item"
        ],
        "title": "EvaluationErrorRow"
      },
      "EvaluationErrorsResponse": {
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/EvaluationErrorRow"
            },
            "type": "array",
            "title": "Errors"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "perPage": {
            "type": "integer",
            "title": "Perpage"
          }
        },
        "type": "object",
        "required": [
          "errors",
          "total",
          "page",
          "perPage"
        ],
        "title": "EvaluationErrorsResponse"
      },
...
      "EvaluationResponse": {
        "properties": {
          "evaluation": {
            "$ref": "#/components/schemas/EvaluationOut"
          }
        },
        "type": "object",
        "required": [
          "evaluation"
        ],
        "title": "EvaluationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetricRow": {
        "properties": {
          "trainingRunId": {
            "type": "string",
            "format": "uuid",
            "title": "Trainingrunid"
          },
          "epoch": {
            "type": "integer",
            "title": "Epoch"
          },
          "split": {
            "type": "string",
            "enum": [
              "train",
              "test",
              "validation"
            ],
            "title": "Split"
          },
          "metricName": {
            "type": "string",
            "title": "Metricname"
          },
          "metricValue": {
            "type": "number",
            "title": "Metricvalue"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "trainingRunId",
          "epoch",
          "split",
          "metricName",
          "metricValue",
          "createdAt"
        ],
        "title": "MetricRow"
      },
...
      "RunCreatedResponse": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunRow"
          }
        },
        "type": "object",
        "required": [
          "run"
        ],
        "title": "RunCreatedResponse"
      },
      "RunDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "canceled"
            ],
            "title": "Status"
          },
          "hyperparameters": {
            "title": "Hyperparameters"
          },
          "failedMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failedmessage"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "datasetVersion": {
            "$ref": "#/components/schemas/RunVersion"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricRow"
            },
            "type": "array",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "hyperparameters",
          "failedMessage",
          "startedAt",
          "completedAt",
          "createdAt",
          "datasetVersion",
          "metrics"
        ],
        "title": "RunDetail"
      },
      "RunDetailResponse": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunDetail"
          }
        },
        "type": "object",
        "required": [
          "run"
        ],
        "title": "RunDetailResponse"
      },
      "RunListResponse": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Runs"
          }
        },
        "type": "object",
        "required": [
          "runs"
        ],
        "title": "RunListResponse"
      },
...
      "RunStatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "canceled"
            ],
            "title": "Status"
          },
          "epochsTotal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Epochstotal"
          },
          "latestMetrics": {
            "items": {
              "$ref": "#/components/schemas/MetricRow"
            },
            "type": "array",
            "title": "Latestmetrics"
          }
        },
        "type": "object",
        "required": [
          "status",
          "epochsTotal",
          "latestMetrics"
        ],
        "title": "RunStatusResponse"
      },
      "RunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "canceled"
            ],
            "title": "Status"
          },
          "failedMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failedmessage"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "datasetVersionId": {
            "type": "string",
            "format": "uuid",
            "title": "Datasetversionid"
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationBrief"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "failedMessage",
          "completedAt",
          "startedAt",
          "createdAt",
          "datasetVersionId"
        ],
        "title": "RunSummary"
      },
...
      "TrainBody": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "datasetVersionId": {
            "type": "string",
            "format": "uuid",
            "title": "Datasetversionid"
          },
          "hyperparameters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrainerSelections"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "datasetVersionId"
        ],
        "title": "TrainBody"
      },
      "TrainerSelections": {
        "properties": {
          "epochs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Epochs"
          },
          "batchSize": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string",
                "const": "auto"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batchsize"
          },
          "learningRate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learningrate"
          },
          "earlyStopPatience": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earlystoppatience"
          },
          "encoderId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encoderid"
          },
          "useClassWeights": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Useclassweights"
          },
          "augmentations": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Augmentations"
          },
          "imageSize": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imagesize"
          },
          "validationMetric": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validationmetric"
          },
          "optimizer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimizer"
          }
        },
        "type": "object",
        "title": "TrainerSelections",
        "description": "User-facing hyperparameter choices (camelCase on the wire, and in runs.hyperparameters)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}